2
Update SemanticsObjectTest.mm

2

Update semantics_node.h

make semanticsflags a struct

+++ engine/src/flutter/shell/platform/android/io/flutter/view/AccessibilityBridge.java
@@ -2514,7 +2514,7 @@ public class AccessibilityBridge extends AccessibilityNodeProvider {
       previousScrollExtentMax = scrollExtentMax;
       previousScrollExtentMin = scrollExtentMin;
 
-      flags = buffer.getInt();
+      flags = buffer.getLong();
       actions = buffer.getInt();
       maxValueLength = buffer.getInt();
       currentValueLength = buffer.getInt();
